refactor: extract method to combine permission into utility class
