+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/handler/saas/ClientMigrationHandler.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.migration.identity.handler.saas;
 
+import static io.camunda.migration.identity.MigrationUtil.extractCombinedPermissions;
 import static io.camunda.migration.identity.config.saas.StaticEntities.getOperateClientPermissions;
 import static io.camunda.migration.identity.config.saas.StaticEntities.getTasklistClientPermissions;
 import static io.camunda.migration.identity.config.saas.StaticEntities.getZeebeClientPermissions;
@@ -20,15 +21,8 @@ import io.camunda.migration.identity.handler.MigrationHandler;
 import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.service.AuthorizationServices;
 import io.camunda.service.AuthorizationServices.CreateAuthorizationRequest;
-import io.camunda.zeebe.protocol.record.value.AuthorizationOwnerType;
-import io.camunda.zeebe.protocol.record.value.AuthorizationResourceType;
-import io.camunda.zeebe.protocol.record.value.PermissionType;
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
@@ -105,15 +99,6 @@ public class ClientMigrationHandler extends MigrationHandler<Members> {
   public static List<CreateAuthorizationRequest> getCombinedPermissions(
       final String clientId, final List<Permission> clientTypes) {
 
-    // Key used to group permission sets
-    record PermissionKey(
-        String ownerId,
-        AuthorizationOwnerType ownerType,
-        String resourceId,
-        AuthorizationResourceType resourceType) {}
-
-    final Map<PermissionKey, Set<PermissionType>> permissionMap = new HashMap<>();
-
     final List<CreateAuthorizationRequest> allPermissions = new ArrayList<>();
 
     if (clientTypes.contains(Permission.ZEEBE)) {
@@ -126,29 +111,6 @@ public class ClientMigrationHandler extends MigrationHandler<Members> {
       allPermissions.addAll(getTasklistClientPermissions(clientId));
     }
 
-    for (final CreateAuthorizationRequest request : allPermissions) {
-      final PermissionKey key =
-          new PermissionKey(
-              request.ownerId(), request.ownerType(), request.resourceId(), request.resourceType());
-
-      permissionMap.merge(
-          key,
-          new HashSet<>(request.permissionTypes()),
-          (existing, incoming) -> {
-            existing.addAll(incoming);
-            return existing;
-          });
-    }
-
-    return permissionMap.entrySet().stream()
-        .map(
-            entry ->
-                new CreateAuthorizationRequest(
-                    entry.getKey().ownerId(),
-                    entry.getKey().ownerType(),
-                    entry.getKey().resourceId(),
-                    entry.getKey().resourceType(),
-                    entry.getValue()))
-        .toList();
+    return extractCombinedPermissions(allPermissions);
   }
 }
